Methods - Accessing fields

Index

EnergyModelsHydrogen.current_op

EnergyModelsHydrogen.degradation_rate

EnergyModelsHydrogen.discharge_charge

EnergyModelsHydrogen.electricity_resource

EnergyModelsHydrogen.is_last

EnergyModelsHydrogen.last_op

EnergyModelsHydrogen.level_charge

EnergyModelsHydrogen.max_load

EnergyModelsHydrogen.min_load

EnergyModelsHydrogen.op_per

EnergyModelsHydrogen.opex_off

EnergyModelsHydrogen.opex_shutdown

EnergyModelsHydrogen.opex_startup

EnergyModelsHydrogen.opex_state

EnergyModelsHydrogen.p_charge

EnergyModelsHydrogen.p_max

EnergyModelsHydrogen.p_min

EnergyModelsHydrogen.prev_op

EnergyModelsHydrogen.ramp_down

EnergyModelsHydrogen.ramp_limit

EnergyModelsHydrogen.ramp_up

EnergyModelsHydrogen.stack_lifetime

EnergyModelsHydrogen.stack_replacement_cost

EnergyModelsHydrogen.strat_per

EnergyModelsHydrogen.time_off

EnergyModelsHydrogen.time_shutdown

EnergyModelsHydrogen.time_startup

EnergyModelsHydrogen.time_state

TimeStruct.strat_periods

AbstractElectrolyzer types

EnergyModelsHydrogen.degradation_rate — Function

degradation_rate(n::Electrolyzer)

Returns the degradation rate of electrolyzer n.

EnergyModelsHydrogen.stack_replacement_cost — Function

stack_replacement_cost(n::Electrolyzer)
stack_replacement_cost(n::Electrolyzer, t_inv)

Returns the stack replacement costs of electrolyzer n as TimeProfile or in strategic period t_inv.

EnergyModelsHydrogen.stack_lifetime — Function

stack_lifetime(n::Electrolyzer)

Returns the stack lfetime of electrolyzer n.

AbstractReformer types

EnergyModelsHydrogen.opex_startup — Function

opex_startup(n::AbstractReformer)
opex_startup(n::AbstractReformer, t)

Returns the startup OPEX of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.opex_shutdown — Function

opex_shutdown(n::AbstractReformer)
opex_shutdown(n::AbstractReformer, t)

Returns the shutdown OPEX of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.opex_off — Function

opex_off(n::AbstractReformer)
opex_off(n::AbstractReformer, t)

Returns the offline OPEX of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.time_startup — Function

time_startup(n::AbstractReformer)
time_startup(n::AbstractReformer, t)

Returns the minimum startup time of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.time_shutdown — Function

time_shutdown(n::AbstractReformer)
time_shutdown(n::AbstractReformer, t)

Returns the minimum shutdown time of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.time_off — Function

time_off(n::AbstractReformer)
time_off(n::AbstractReformer, t)

Returns the minimum offline time of AbstractReformer n as TimeProfile or in in operational period t.

EnergyModelsHydrogen.ramp_limit — Function

ramp_limit(n::AbstractReformer)

Returns the AbstractRampParameters type of AbstractReformer n.

AbstractH2Storage types

EnergyModelsHydrogen.discharge_charge — Function

discharge_charge(n::AbstractH2Storage)

Returns the discharge to charge ratio of AbstractH2Storage n.

EnergyModelsHydrogen.level_charge — Function

level_charge(n::AbstractH2Storage)

Returns the level to charge ratio of AbstractH2Storage n.

HydrogenStorage types

EnergyModelsHydrogen.p_charge — Function

p_charge(n::HydrogenStorage)

Returns the charging pressure of HydrogenStorage, n. This pressure corresponds to the inlet pressure of the node, that is the pressure used for calculating the energy demand.

EnergyModelsHydrogen.p_min — Function

p_min(n::HydrogenStorage)

Returns the minimum pressure of HydrogenStorage, n. This pressure corresponds to an empty storage.

EnergyModelsHydrogen.p_max — Function

p_max(n::HydrogenStorage)

Returns the maximum pressure of HydrogenStorage, n. This pressure corresponds to a full storage.

EnergyModelsHydrogen.electricity_resource — Function

el_res(n::HydrogenStorage)

Returns the resource of HydrogenStorage, n which corresponds to electricity in the system.

LoadLimit and Node types

EnergyModelsHydrogen.min_load — Function

min_load(load_lim::AbstractLoadLimits)
min_load(load_lim::AbstractLoadLimits, t)

Returns the minimum load of AbstractLoadLimits load_lim as TimeProfile or in operational period t.

Note

The default LoadLimits does not allow for time dependent load limits. In this case, the function returns a FixedProfile of the provided value.

min_load(n::EMB.Node)
min_load(n::EMB.Node, t)

Returns the minimum load of Node n as TimeProfile or in operational period t.

EnergyModelsHydrogen.max_load — Function

max_load(load_lim::AbstractLoadLimits)
min_load(load_lim::AbstractLoadLimits, t)

Returns the maximum load of AbstractLoadLimits load_lim as TimeProfile or in operational period t.

Note

The default LoadLimits does not allow for time dependent load limits. In this case, the function returns a FixedProfile of the provided value.

max_load(n::EMB.Node)
max_load(n::EMB.Node, t)

Returns the maximum load of Node n.

AbstractRampParameters and AbstractReformer types

EnergyModelsHydrogen.ramp_up — Function

ramp_up(ramp_param::UnionRampUp)
ramp_up(ramp_param::UnionRampUp, t)

Returns the maximum positive rate of change of UnionRampUp ramp_param as TimeProfile or in operational period t.

ramp_up(n::AbstractReformer)
ramp_up(n::AbstractReformer, t)

Returns the maximum positive rate of change of AbstractReformer n as TimeProfile or in operational period t.

EnergyModelsHydrogen.ramp_down — Function

ramp_down(ramp_param::UnionRampDown)
ramp_down(ramp_param::UnionRampDown, t)

Returns the maximum negative rate of change of UnionRampDown ramp_param as TimeProfile or in operational period t.

ramp_down(n::AbstractReformer)
ramp_down(n::AbstractReformer, t)

Returns the maximum negative rate of change of AbstractReformer n as TimeProfile or in operational period t.

CommitParameters types

EnergyModelsHydrogen.opex_state — Function

opex_state(com_par::CommitParameters)
opex_state(com_par::CommitParameters, t)

Returns the unit commitment OPEX as TimeProfile or in operational period t.

EnergyModelsHydrogen.time_state — Function

time_state(com_par::CommitParameters)
time_state(com_par::CommitParameters, t)

Returns the minimum time in the state as TimeProfile or in operational period t.

ElecPeriods types

TimeStruct.strat_periods — Function

TS.strat_periods(pers::ElecPeriods)

Returns the strategic periods of an ElecPeriods pers.

EnergyModelsHydrogen.strat_per — Function

strat_per(pers::ElecPeriods)

Returns the current strategic period of an ElecPeriods pers.

EnergyModelsHydrogen.op_per — Function

op_per(pers::ElecPeriods)

Returns the current operational period of an ElecPeriods pers.

EnergyModelsHydrogen.is_last — Function

is_last(pers::ElecPeriods)

Boolean indicator whether the representative period or operational scenario is the last within a strategic period of an ElecPeriods pers.

RefPeriods types

EnergyModelsHydrogen.prev_op — Function

prev_op(pers::RefPeriods)

Returns the previous operational period of a RefPeriods pers.

When the previous operational period is nothing, it returns the last operational period within the given time structure.

EnergyModelsHydrogen.current_op — Function

current_op(pers::RefPeriods)

Returns the current operational period of a RefPeriods pers.

EnergyModelsHydrogen.last_op — Function

last_op(pers::RefPeriods)

Returns the last operational period of a RefPeriods pers.