Types

Index

EnergyModelsHydrogen.AbstractElectrolyzer

EnergyModelsHydrogen.AbstractH2Storage

EnergyModelsHydrogen.AbstractHydrogenNetworkNode

EnergyModelsHydrogen.AbstractLoadLimits

EnergyModelsHydrogen.AbstractRampParameters

EnergyModelsHydrogen.AbstractReformer

EnergyModelsHydrogen.ElecPeriods

EnergyModelsHydrogen.RefPeriods

Nodal supertypes

EnergyModelsHydrogen.AbstractHydrogenNetworkNode — Type

Abstract supertype for all hydrogen network nodes.

EnergyModelsHydrogen.AbstractElectrolyzer — Type

AbstractElectrolyzer <: AbstractHydrogenNetworkNode

Abstract supertype for all electrolyzer nodes.

EnergyModelsHydrogen.AbstractReformer — Type

Abstract supertype for all reformer nodes.

EnergyModelsHydrogen.AbstractH2Storage — Type

AbstractH2Storage{T} <: Storage{T}

Abstract type for different implementations of hydrogen storage nodes.

Parameter supertypes

EnergyModelsHydrogen.AbstractLoadLimits — Type

AbstractLoadLimits{T}

Abstract type for the load limits. This type can be used to incorporate other types for the load limit.

EnergyModelsHydrogen.AbstractRampParameters — Type

AbstractRampParameters

Abstract type for different ramp parameter configurations.

Auxiliary types

EnergyModelsHydrogen.ElecPeriods — Type

ElecPeriods

Contains information for calculating the constraints for AbstractElectrolyzer node types.

Fields

sps::TS.AbstractStratPers are the strategic periods of the TimeStructure.

sp::TS.AbstractStrategicPeriod is the current strategic period.

op::TS.AbstractStrategicPeriod is the current operational period.

last::Bool is a boolean indicator of the last period. It is used for calculating the bounds for the last operational periods within a strategic period.

EnergyModelsHydrogen.RefPeriods — Type

RefPeriods{S<:Union{TS.OperationalPeriod, Nothing}}

Contains information for calculating the constraints for Reformer node types.

Fields

previous::S is the previous operational period received from the withprev iterator

current::TS.OperationalPeriod is the current operational period.

last::TS.OperationalPeriod is the last operational period in the current SimpleTimes structure within a strategic period (or representative period, if present, or operational scenario, if present).